Cluster and Array Response

+++ Environment.go
@@ -1,7 +1,7 @@
 package main
 
 type Environment struct {
-	length int
-	width  int
-	height int
+	length float64
+	width  float64
+	height float64
 }
